fix: add back reverting to ejector and rename intake and ejector files
