+++ fission/src/ui/panels/configuring/assembly-config/interfaces/ConfigureGamepieceIntakeInterface.tsx
@@ -3,6 +3,7 @@ import { Stack } from "@mui/material"
 import { useCallback, useEffect, useMemo, useRef, useState } from "react"
 import * as THREE from "three"
 import SelectButton from "@/components/SelectButton"
+import EjectableSceneObject from "@/mirabuf/EjectableSceneObject"
 import type { RigidNodeId } from "@/mirabuf/MirabufParser"
 import type MirabufSceneObject from "@/mirabuf/MirabufSceneObject"
 import type { RigidNodeAssociate } from "@/mirabuf/MirabufSceneObject"
@@ -10,8 +11,9 @@ import EventSystem from "@/systems/EventSystem.ts"
 import { PAUSE_REF_ASSEMBLY_CONFIG } from "@/systems/physics/PhysicsTypes"
 import type GizmoSceneObject from "@/systems/scene/GizmoSceneObject"
 import World from "@/systems/World"
+import Checkbox from "@/ui/components/Checkbox"
 import StatefulSlider from "@/ui/components/StatefulSlider"
-import { Button, LabelWithTooltip, Spacer, ToggleButton, ToggleButtonGroup } from "@/ui/components/StyledComponents"
+import { Button, Spacer } from "@/ui/components/StyledComponents"
 import TransformGizmoControl from "@/ui/components/TransformGizmoControl"
 import {
     convertArrayToThreeMatrix4,
@@ -22,8 +24,11 @@ import {
 import type { ConfigurationSubpanelComponent } from "@/panels/configuring/assembly-config/ConfigTypes.ts"
 
 // slider constants
-const MIN_VELOCITY = 0.0
-const MAX_VELOCITY = 20.0
+const MIN_ZONE_SIZE = 0.1
+const MAX_ZONE_SIZE = 1.0
+const MIN_ANIMATION_DURATION = 0.1
+const MAX_ANIMATION_DURATION = 2.0
+const ANIMATION_DURATION_STEP = 0.05
 
 /**
  * Saves ejector configuration to selected robot.
@@ -49,13 +54,15 @@ const MAX_VELOCITY = 20.0
  * @param selectedNode Selected node that configuration is relative to.
  */
 function save(
-    ejectorVelocity: number,
+    zoneSize: number,
     gizmo: GizmoSceneObject,
     selectedRobot: MirabufSceneObject,
     selectedNode?: RigidNodeId,
-    ejectOrder?: "FIFO" | "LIFO"
+    showZoneAlways?: boolean,
+    maxPieces?: number,
+    animationDuration?: number
 ) {
-    if (!selectedRobot?.ejectorPreferences || !gizmo) {
+    if (!selectedRobot?.intakePreferences || !gizmo) {
         return
     }
 
@@ -66,65 +73,91 @@ function save(
         return
     }
 
-    const gizmoTransformation = gizmo.obj.matrixWorld
+    const translation = new THREE.Vector3(0, 0, 0)
+    const rotation = new THREE.Quaternion(0, 0, 0, 1)
+    gizmo.obj.matrixWorld.decompose(translation, rotation, new THREE.Vector3(1, 1, 1))
+
+    const gizmoTransformation = new THREE.Matrix4().compose(translation, rotation, new THREE.Vector3(1, 1, 1))
     const robotTransformation = convertJoltMat44ToThreeMatrix4(
         World.physicsSystem.getBody(nodeBodyId)!.GetWorldTransform()
     )
     const deltaTransformation = gizmoTransformation.premultiply(robotTransformation.invert())
 
-    selectedRobot.ejectorPreferences.deltaTransformation = convertThreeMatrix4ToArray(deltaTransformation)
-    selectedRobot.ejectorPreferences.parentNode = selectedNode
-    selectedRobot.ejectorPreferences.ejectorVelocity = ejectorVelocity
+    selectedRobot.intakePreferences.deltaTransformation = convertThreeMatrix4ToArray(deltaTransformation)
+    selectedRobot.intakePreferences.parentNode = selectedNode
+    selectedRobot.intakePreferences.zoneDiameter = zoneSize
+    if (showZoneAlways !== undefined) {
+        selectedRobot.intakePreferences.showZoneAlways = showZoneAlways
+    }
 
-    selectedRobot.ejectorPreferences.ejectOrder = ejectOrder!
-
+    selectedRobot.intakePreferences.maxPieces = maxPieces!
+    selectedRobot.intakePreferences.animationDuration = animationDuration!
     selectedRobot.savePreferences()
 }
 
-const ConfigureShotTrajectoryInterface: ConfigurationSubpanelComponent = ({ selectedAssembly }) => {
+const ConfigureGamepieceIntakeInterface: ConfigurationSubpanelComponent = ({
+    selectedAssembly,
+    registerCleanupFunction,
+}) => {
     const [selectedNode, setSelectedNode] = useState<RigidNodeId | undefined>(undefined)
-    const [ejectorVelocity, setEjectorVelocity] = useState<number>((MIN_VELOCITY + MAX_VELOCITY) / 2.0)
-    const [ejectOrder, setEjectOrder] = useState<"FIFO" | "LIFO">(
-        selectedAssembly.ejectorPreferences?.ejectOrder || "FIFO"
+    const [zoneSize, setZoneSize] = useState<number>((MIN_ZONE_SIZE + MAX_ZONE_SIZE) / 2.0)
+    const [showZoneAlways, setShowZoneAlways] = useState<boolean>(false)
+    const [maxPieces, setMaxPieces] = useState<number>(selectedAssembly.intakePreferences?.maxPieces || 1)
+    const [animationDuration, setAnimationDuration] = useState<number>(
+        selectedAssembly.intakePreferences?.animationDuration || 0.5
     )
 
     const gizmoRef = useRef<GizmoSceneObject | undefined>(undefined)
 
     const saveEvent = useCallback(() => {
         if (gizmoRef.current && selectedAssembly) {
-            save(ejectorVelocity, gizmoRef.current, selectedAssembly, selectedNode, ejectOrder)
-            const currentGp = selectedAssembly.activeEjectables[0]
-            selectedAssembly.setEjectable(undefined)
-            selectedAssembly.setEjectable(currentGp)
+            save(
+                zoneSize,
+                gizmoRef.current,
+                selectedAssembly,
+                selectedNode,
+                showZoneAlways,
+                maxPieces,
+                animationDuration
+            )
+            selectedAssembly.updateIntakeSensor()
         }
-    }, [selectedAssembly, selectedNode, ejectorVelocity, ejectOrder])
+    }, [selectedAssembly, selectedNode, zoneSize, showZoneAlways, maxPieces, animationDuration])
 
     useEffect(() => {
         return EventSystem.listen("ConfigurationSavedEvent", saveEvent)
     }, [saveEvent])
 
+    useEffect(() => {
+        if (!gizmoRef.current) {
+            return
+        }
+
+        gizmoRef.current.obj.scale.set(zoneSize, zoneSize, zoneSize)
+    }, [zoneSize])
+
     const placeholderMesh = useMemo(() => {
-        return new THREE.Mesh(
-            new THREE.ConeGeometry(0.1, 0.4, 4).rotateX(Math.PI / 2.0).translate(0, 0, 0.2),
-            // TODO: dynamic color
-            World.sceneRenderer.createToonMaterial(
-                convertReactRgbaColorToThreeColor({ r: 255, g: 255, b: 255, a: 255 })
-            )
+        // TODO: dynamic color?
+        const material = World.sceneRenderer.createToonMaterial(
+            convertReactRgbaColorToThreeColor({ r: 255, g: 255, b: 255, a: 255 })
         )
+        material.transparent = true
+        material.opacity = 0.6
+        return new THREE.Mesh(new THREE.SphereGeometry(0.5), material)
     }, [])
 
     const gizmoComponent = useMemo(() => {
-        if (selectedAssembly?.ejectorPreferences) {
+        if (selectedAssembly?.intakePreferences) {
             const postGizmoCreation = (gizmo: GizmoSceneObject) => {
                 const material = (gizmo.obj as THREE.Mesh).material as THREE.Material
                 material.depthTest = false
 
                 const deltaTransformation = convertArrayToThreeMatrix4(
-                    selectedAssembly.ejectorPreferences!.deltaTransformation
+                    selectedAssembly.intakePreferences!.deltaTransformation
                 )
 
                 let nodeBodyId = selectedAssembly.mechanism.nodeToBody.get(
-                    selectedAssembly.ejectorPreferences!.parentNode ?? selectedAssembly.rootNodeId
+                    selectedAssembly.intakePreferences!.parentNode ?? selectedAssembly.rootNodeId
                 )
                 if (!nodeBodyId) {
                     // In the event that something about the id generation for the rigid nodes changes and parent node id is no longer in use
@@ -137,18 +170,18 @@ const ConfigureShotTrajectoryInterface: ConfigurationSubpanelComponent = ({ sele
                 )
                 const gizmoTransformation = deltaTransformation.premultiply(robotTransformation)
 
-                gizmo.obj.position.setFromMatrixPosition(gizmoTransformation)
-                gizmo.obj.rotation.setFromRotationMatrix(gizmoTransformation)
+                gizmo.setTransform(gizmoTransformation)
             }
 
             return (
                 <TransformGizmoControl
-                    key="shot-transform-gizmo"
+                    key="pickup-transform-gizmo"
                     size={1.5}
                     gizmoRef={gizmoRef}
                     defaultMode="translate"
                     defaultMesh={placeholderMesh}
                     scaleDisabled={true}
+                    rotateDisabled={true}
                     postGizmoCreation={postGizmoCreation}
                 />
             )
@@ -157,29 +190,43 @@ const ConfigureShotTrajectoryInterface: ConfigurationSubpanelComponent = ({ sele
             return <></>
         }
     }, [
+        selectedAssembly.intakePreferences,
         placeholderMesh,
-        selectedAssembly.ejectorPreferences,
-        selectedAssembly.mechanism.nodeToBody.get,
         selectedAssembly.rootNodeId,
+        selectedAssembly.mechanism.nodeToBody,
     ])
 
     useEffect(() => {
-        if (selectedAssembly?.ejectorPreferences) {
-            setEjectorVelocity(selectedAssembly.ejectorPreferences.ejectorVelocity)
-            setSelectedNode(selectedAssembly.ejectorPreferences.parentNode)
-            setEjectOrder(selectedAssembly.ejectorPreferences.ejectOrder)
+        if (selectedAssembly?.intakePreferences) {
+            setZoneSize(selectedAssembly.intakePreferences.zoneDiameter)
+            setSelectedNode(selectedAssembly.intakePreferences.parentNode)
+            setMaxPieces(selectedAssembly.intakePreferences.maxPieces)
+            setShowZoneAlways(selectedAssembly.intakePreferences.showZoneAlways ?? false)
+            setAnimationDuration(selectedAssembly.intakePreferences.animationDuration ?? 0.5)
         } else {
             setSelectedNode(undefined)
+            setShowZoneAlways(false)
+            setAnimationDuration(0.5)
         }
     }, [selectedAssembly])
 
     useEffect(() => {
         World.physicsSystem.holdPause(PAUSE_REF_ASSEMBLY_CONFIG)
 
+        // Hide the visual indicator when entering configuration mode
+        if (selectedAssembly) {
+            selectedAssembly.setIntakeVisualIndicatorVisible(false)
+        }
+
         return () => {
             World.physicsSystem.releasePause(PAUSE_REF_ASSEMBLY_CONFIG)
+
+            // Show the visual indicator when exiting configuration mode
+            if (selectedAssembly) {
+                selectedAssembly.setIntakeVisualIndicatorVisible(true)
+            }
         }
-    }, [])
+    }, [selectedAssembly])
 
     const trySetSelectedNode = useCallback(
         (body: Jolt.BodyID) => {
@@ -188,7 +235,7 @@ const ConfigureShotTrajectoryInterface: ConfigurationSubpanelComponent = ({ sele
             }
 
             const assoc = World.physicsSystem.getBodyAssociation(body) as RigidNodeAssociate
-            if (!assoc || !assoc.sceneObject || assoc.sceneObject !== selectedAssembly) {
+            if (!assoc || !assoc.sceneObject || assoc.sceneObject != selectedAssembly) {
                 return false
             }
 
@@ -198,8 +245,16 @@ const ConfigureShotTrajectoryInterface: ConfigurationSubpanelComponent = ({ sele
         [selectedAssembly]
     )
 
+    useEffect(() => {
+        const originalPrefs = structuredClone(selectedAssembly.intakePreferences)
+        registerCleanupFunction(undefined, () => {
+            selectedAssembly.intakePreferences = originalPrefs
+            selectedAssembly.updateIntakeSensor()
+        })
+    }, [registerCleanupFunction, selectedAssembly])
+
     return (
-        <>
+        <Stack direction="column">
             {/* Button for user to select the parent node */}
             <SelectButton
                 placeholder="Select parent node"
@@ -207,37 +262,46 @@ const ConfigureShotTrajectoryInterface: ConfigurationSubpanelComponent = ({ sele
                 onSelect={(body: Jolt.Body) => trySetSelectedNode(body.GetID())}
             />
 
-            {/* Toggle for adjusting eject order */}
-            <Stack direction="row" spacing={2} alignItems="center" className="mt-4">
-                <LabelWithTooltip
-                    labelText="Eject Order"
-                    tooltipText="Choose how to eject pieces: FIFO (first in, first out) ejects the oldest-loaded item first, or LIFO (last in, first out) ejects the most recently loaded item first."
-                />
-                <ToggleButtonGroup
-                    value={ejectOrder}
-                    exclusive
-                    onChange={(_: unknown, v: "FIFO" | "LIFO") => v && setEjectOrder(v)}
-                >
-                    <ToggleButton value="FIFO">FIFO</ToggleButton>
-                    <ToggleButton value="LIFO">LIFO</ToggleButton>
-                </ToggleButtonGroup>
-            </Stack>
-
             {/* Slider for user to set velocity of ejector configuration */}
             <StatefulSlider
-                label="Velocity"
-                min={MIN_VELOCITY}
-                max={MAX_VELOCITY}
-                defaultValue={ejectorVelocity}
+                label="Intake Zone Diameter (m)"
+                min={MIN_ZONE_SIZE}
+                max={MAX_ZONE_SIZE}
+                defaultValue={zoneSize}
                 // TODO:
                 // format={{ minimumFractionDigits: 2, maximumFractionDigits: 2 }}
                 onChange={vel => {
-                    setEjectorVelocity(vel as number)
+                    setZoneSize(vel as number)
                 }}
                 step={0.01}
             />
+            <StatefulSlider
+                label="Intake Animation Duration (s)"
+                min={MIN_ANIMATION_DURATION}
+                max={MAX_ANIMATION_DURATION}
+                defaultValue={animationDuration ?? 0.5}
+                onChange={v => {
+                    const val = typeof v === "number" ? v : v[0]
+                    setAnimationDuration(val)
+                    EjectableSceneObject.setAnimationDuration(val)
+                }}
+                step={ANIMATION_DURATION_STEP}
+                // TODO:
+                // format={{ maximumFractionDigits: 2 }}
+            />
 
-            <Spacer height={10} />
+            {/* Slider for adjusting max pieces the robot can intake */}
+            <StatefulSlider
+                label="Max Pieces"
+                min={1}
+                max={50}
+                step={1}
+                defaultValue={maxPieces ?? 1}
+                onChange={v => setMaxPieces(v as number)}
+            />
+
+            {/* Checkbox for showing intake zone indicator at all times */}
+            <Checkbox label="Show intake zone indicator always" checked={showZoneAlways} onClick={setShowZoneAlways} />
             {gizmoComponent}
             <Spacer height={10} />
             <Button
@@ -249,15 +313,16 @@ const ConfigureShotTrajectoryInterface: ConfigurationSubpanelComponent = ({ sele
                         gizmoRef.current.obj.position.setFromMatrixPosition(robotTransformation)
                         gizmoRef.current.obj.rotation.setFromRotationMatrix(robotTransformation)
                     }
-                    setEjectorVelocity(1)
+                    setZoneSize(0.5)
                     setSelectedNode(selectedAssembly?.rootNodeId)
-                    setEjectOrder(selectedAssembly.ejectorPreferences?.ejectOrder ?? "FIFO")
+                    setMaxPieces(selectedAssembly.intakePreferences?.maxPieces ?? 1)
+                    setAnimationDuration(0.5)
                 }}
             >
                 Reset
             </Button>
-        </>
+        </Stack>
     )
 }
 
-export default ConfigureShotTrajectoryInterface
+export default ConfigureGamepieceIntakeInterface
